+++ src/Trarizon.Library/Collections/AutoAlloc/AutoAllocList.cs
@@ -1,15 +1,15 @@
 ﻿using Trarizon.Library.Buffers;
 
 namespace Trarizon.Library.Collections.AutoAlloc;
-public class AutoAllocList<T> : BaseAutoAllocList<T> where T : class
+public class AutoAllocList<T>(IObjectAllocator<T> allocator) :
+    AutoAllocList<T, IObjectAllocator<T>>(allocator)
+    where T : class
+{ }
+
+public class AutoAllocList<T, TAllocator>(TAllocator allocator) : BaseAutoAllocList<T>
+    where T : class
+    where TAllocator : IObjectAllocator<T>
 {
-    private readonly IObjectAllocator<T> _allocator;
-
-    public AutoAllocList(IObjectAllocator<T> allocator)
-    {
-        _allocator = allocator;
-    }
-
-    protected override T Allocate() => _allocator.Allocate();
-    protected override void Release(T item) => _allocator.Release(item);
+    protected override T Allocate() => allocator.Allocate();
+    protected override void Release(T item) => allocator.Release(item);
 }