+++ src/Trarizon.Library/Collections/TraSpan/Views.cs
@@ -4,13 +4,17 @@ using Trarizon.Library.Collections.StackAlloc;
 namespace Trarizon.Library.Collections;
 public static partial class TraSpan
 {
-#if !NETSTANDARD
-    
     public static ReversedSpan<T> AsReversed<T>(this Span<T> span)
+#if NETSTANDARD
+        => new(span);
+#else
         => new(ref span.DangerousGetReferenceAt(span.Length - 1), span.Length);
+#endif
 
     public static ReadOnlyReversedSpan<T> AsReversed<T>(this ReadOnlySpan<T> span)
+#if NETSTANDARD
+        => new(span);
+#else
         => new(in span.DangerousGetReferenceAt(span.Length - 1), span.Length);
-
 #endif
 }